error implementation added classes fixed accordingly.

+++ data/src/main/java/com/app/signal/data/room/dao/AnyDao.kt
@@ -8,12 +8,12 @@ abstract class AnyDao<T> {
     abstract suspend fun insert(obj: T): Long
 
     @Update
-    abstract suspend fun update(obj: T): Long
+    abstract suspend fun update(obj: T): Int
 
     @Transaction
     open suspend fun insertOrUpdate(obj: T): Long {
         val id = insert(obj)
-        return if (id == -1L) update(obj) else id
+        return if (id == -1L) update(obj).toLong() else id
     }
 
 }
